RANN-package {RANN}	R Documentation

Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library

Description

Finds the k nearest neighbours for every point in a given dataset in O(N log N) time using Arya and Mount's ANN library (v1.1.3). There is support for approximate as well as exact searches, fixed radius searches and 'bd' as well as 'kd' trees. The distance is computed using the L2 (Euclidean) metric. Please see package 'RANN.L1' for the same functionality using the L1 (Manhattan, taxicab) metric.

Author(s)

Maintainer: Gregory Jefferis jefferis@gmail.com (ORCID)

Authors:

Samuel E. Kemp

Sunil Arya (ORCID) [copyright holder]

David Mount (ORCID) [copyright holder]

Other contributors:

Kirill Müller (ORCID) [contributor]

University of Maryland (ANN library is copyright University of Maryland and Sunil Arya and David Mount. See file COPYRIGHT for details) [copyright holder]

See Also

nn2

[Package RANN version 2.6.2 Index]